coli2DA version 2.2
New parameters to functions 'tile' and 'tileCross': mapX, mapY. Both are optional and it should used for to specify a adjust value, for to calculate the object position, if the 'map' function was called with its parameters different of zero. E.g.: 'map( 30, 60 )'. more information abount (comming soon) in the Wiki.